Move `approximately_equal` to `utils` and share across tests
